Разделяет reducer, actionType и actionCreator. Удаляет функцию для генерации моков из е2е-теста карты
