+++ src/action-type.js
@@ -1,0 +1,8 @@
+const ActionType = {
+  CHANGE_CITY: `CHANGE_CITY`,
+  SELECT_OFFERS: `SELECT_OFFERS`,
+  CHANGE_ACTIVE_CARD: `CHANGE_ACTIVE_CARD`,
+  CHANGE_SCREEN_TYPE: `CHANGE_SCREEN_TYPE`
+};
+
+export default ActionType;
