+++ src/reducer.js
@@ -1,6 +1,7 @@
 import {offerMocks} from './mocks/offers.js';
 import {getCities, getOffers} from './util.js';
 import {ScreenType} from './const.js';
+import ActionType from './action-type.js';
 
 const cities = getCities(offerMocks);
 
@@ -11,20 +12,6 @@ const initialState = {
   selectedOffers: getOffers(cities[0], offerMocks),
   activeCard: null,
   screenType: ScreenType.MAIN
-};
-
-const ActionType = {
-  CHANGE_CITY: `CHANGE_CITY`,
-  SELECT_OFFERS: `SELECT_OFFERS`,
-  CHANGE_ACTIVE_CARD: `CHANGE_ACTIVE_CARD`,
-  CHANGE_SCREEN_TYPE: `CHANGE_SCREEN_TYPE`
-};
-
-const ActionCreator = {
-  changeCity: (cityIndex) => ({type: ActionType.CHANGE_CITY, payload: cityIndex}),
-  selectOffers: () => ({type: ActionType.SELECT_OFFERS}),
-  changeActiveCard: (offer) => ({type: ActionType.CHANGE_ACTIVE_CARD, payload: offer}),
-  changeScreenType: (screenType) => ({type: ActionType.CHANGE_SCREEN_TYPE, payload: screenType})
 };
 
 const reducer = (state = initialState, action) => {
@@ -45,4 +32,5 @@ const reducer = (state = initialState, action) => {
   return state;
 };
 
-export {reducer, ActionCreator, ActionType, initialState};
+export default reducer;
+export {initialState};
