+++ src/reducer.test.js
@@ -1,4 +1,5 @@
-import {reducer, ActionCreator, initialState, ActionType} from './reducer.js';
+import reducer, {initialState} from './reducer.js';
+import ActionType from './action-type.js';
 import {ScreenType} from './const.js';
 
 describe(`Test reducer`, () => {
@@ -163,23 +164,5 @@ describe(`Test reducer`, () => {
       activeCard: null,
       screenType: ScreenType.PROPERTY
     });
-  });
-});
-
-describe(`Test ActionCreator`, () => {
-  it(`ActionCreator.changeCity should return a right action`, () => {
-    expect(ActionCreator.changeCity(1)).toEqual({type: ActionType.CHANGE_CITY, payload: 1});
-  });
-
-  it(`ActionCreator.selectOffe should return a right actionrs`, () => {
-    expect(ActionCreator.selectOffers()).toEqual({type: ActionType.SELECT_OFFERS});
-  });
-
-  it(`ActionCreator.changeActi should return a right actionveCard`, () => {
-    expect(ActionCreator.changeActiveCard(`offer`)).toEqual({type: ActionType.CHANGE_ACTIVE_CARD, payload: `offer`});
-  });
-
-  it(`ActionCreator.changeScre should return a right actionenType`, () => {
-    expect(ActionCreator.changeScreenType(ScreenType.MAIN)).toEqual({type: ActionType.CHANGE_SCREEN_TYPE, payload: ScreenType.MAIN});
   });
 });
